requirements/dev.txt
-r common.txt

Flask-Login==0.4.1
Flask-WTF==0.14.2
